Feat/v2 container

+++ init-v2.sh
@@ -19,6 +19,8 @@ CYAN="\e[36m"
 #REVERSE=$(tput smso)
 #UNDERLINE=$(tput smul)
 
+
+
 log () {
     # Display log messages
     # :param: Log level : error, info or debug
@@ -64,24 +66,89 @@ getDistro() {
     echo "$DISTRO"
 }
 
+copyGitBashCompletion() {
+    log info "[copyGitBashCompletion]: On copy le bashrc User"
+    if [[ ! -f ~/.git-bash-completion.sh ]]
+    then
+        cp linuxinit/git-bash-completion.sh ~/.git-bash-completion.sh
+    else
+        log info "[copyUserBashrc][SKIP]: ~/.git-bash-completion.sh exist"
+    fi
+}
+
+copyGitBashPrompt() {
+    log info "[copyGitBashPrompt]: On copy le bashrc User"
+    if [[ ! -f ~/.git-prompt.sh ]]
+    then
+        cp linuxinit/git-prompt.sh ~/.git-prompt.sh
+    else
+        log info "[copyGitBashPrompt][SKIP]: ~/.git-prompt.sh exist"
+    fi
+}
+
+
 copyRootBashrc() {
     log info "[copyRootBashrc]: On copy le bashrc Root"
-    cp linuxinit/bash_aliases_root ~/.bashrc
+    if [[ ! -f ~/.bashrc ]]
+    then
+        cp linuxinit/bash_aliases_root ~/.bashrc
+    else
+        log info "[copyRootBashrc][SKIP]: ~/.bashrc exist - on insere dans bashrc"
+        echo -e "if [ -f ~/.git-bash-completion.sh ]; then\n   . ~/.git-bash-completion.sh\nfi\n\n" >> ~/.bashrc
+        echo -e "if [ -f ~/.git-prompt.sh ]; then\n   . ~/.git-prompt.sh\nfi\n\n" >> ~/.bashrc
+        echo -e "if [ -f ~/.sh_aliases ]; then\n   . ~/.sh_aliases\nfi\n\n" >> ~/.bashrc
+        echo "PS1='${debian_chroot:+($debian_chroot)}\[\033[01;31m\]\u@\h\[\033[00m\] [\D{%T}] \[\033[01;34m\]\w\[\033[00m\]\[\033[36;40m\]$(__git_ps1 "(%s)")\[\033[00m\] > '" >> ~/.bashrc
+    fi
 }
 
 copyUserBashrc() {
     log info "[copyUserBashrc]: On copy le bashrc User"
-    cp linuxinit/bash_aliases_user ~/.bash_aliases
+    if [[ ! -f ~/.bashrc ]]
+    then
+        cp linuxinit/bash_aliases_user ~/.bashrc
+    else
+        log info "[copyUserBashrc][SKIP]: ~/.bashrc exist - on insere dans bashrc"
+    fi
+}
+
+
+copyUserBashAliases() {
+    log info "[copyUserBashAliases]: On copy le bash_aliases User"
+
+    if [[ ! -f ~/.bashrc ]]
+    then
+      log info "[copyUserBashAliases][INIT]: ~/.bashrc doesnt exist we create it"
+        cp linuxinit/bash_aliases_user ~/.bashrc || log info "[copyUserBashAliases][ERROR]: copy failed"
+    else
+        log info "[copyUserBashAliases][SKIP]: ~/.bashrc exist"
+    fi
+
+    if [[ ! -f ~/.bash_aliases ]]
+    then
+        cp linuxinit/bash_aliases_user ~/.bash_aliases
+    else
+        log info "[copyUserBashAliases][SKIP]: ~/.bash_aliases exist"
+    fi
 }
 
 copyShAliases() {
     log info "[copyShAliases]: On copy le Sh_Aliases"
-    cp linuxinit/sh_aliases ~/.sh_aliases
+    if [[ ! -f ~/.sh_aliases ]]
+    then
+        cp linuxinit/sh_aliases ~/.sh_aliases
+    else
+        log info "[copyShAliases][SKIP]: ~/.sh_aliases exist"
+    fi
 }
 
 copyVimrc() {
     log info "[copyVimrc]: On copy le Vimrc"
-    cp linuxinit/vimrc ~/.vimrc
+    if [[ ! -f ~/.vimrc ]]
+    then
+        cp linuxinit/vimrc ~/.vimrc
+    else
+        log info "[copyVimrc][SKIP]: ~/.vimrc exist"
+    fi
 }
 
 copyPsqlRc() {
@@ -133,18 +200,65 @@ createConfigureSshRoot() {
 
 cleanup() {
     log info "[cleanup]: On supprime le repertoire linuxinit"
-    rm -Rf ./linuxinit
+    rm -Rf ./linuxinit f
+}
+
+detectIfInContainer() {
+ local chk
+ chk=$(grep -E '/(lxc|docker)/[[:xdigit:]]{64}' /proc/1/cgroup | wc -l)
+ if [[ $chk -gt 0 ]]
+ then
+   log info "[detectIfInContainer]: On est dans un container"
+	 return 0
+ else
+   log info "[detectIfInContainer]: On est PAS dans un container"
+	 return 1
+ fi
+}
+
+
+getUserByUid() {
+  # Get the username of the Uid parameters
+  # :param: Uid
+  local name
+  name=$(getent passwd $1 | cut -d: -f1)
+  log info "[getUserByUid]: Le user avec l'uid [$1] est: [$name]"
+  echo "$name"
+}
+
+
+insertPS1BashBashrc() {
+  log info "[insertPS1BashBashrc]: Cas particulier on dirait, on force le bash.bashrc"
+
+  echo "########### CUSTOM /etc/bash.bashrc due to no healthy HOME ###########" >> /etc/bash.bashrc
+  echo "PS1='${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\] [\D{%T}] \[\033[01;34m\]\w\[\033[00m\]\[\033[36;40m\]\[\033[00m\] > '" >> /etc/bash.bashrc
+  cat ./linuxinit/sh_aliases >> /etc/bash.bashrc
+#  echo -e "if [ -f ~/.git-bash-completion.sh ]; then\n   . ~/.git-bash-completion.sh\nfi\n\n" >> ~/.bashrc
+#  echo -e "if [ -f ~/.git-prompt.sh ]; then\n   . ~/.git-prompt.sh\nfi\n\n" >> ~/.bashrc
+#  echo -e "if [ -f ~/.sh_aliases ]; then\n   . ~/.sh_aliases\nfi\n\n" >> ~/.bashrc
 }
 
 currentDistro=$(getDistro)
+userUid1000=$(getUserByUid 1000)
 
-log info "---=== Debut de Configuration de becane pour [$USER] [$currentDistro] ===---"
+if [[ -z "$USER" ]] || [[ "$USER" == "" ]]
+then
+    USER=$(whoami)
+    log info "USER Not defined on essaye whoami [$USER]"
+else
+    log info "USER defined [$USER]"
+fi
+
+log info "---=== Debut de Configuration de becane pour User: [$USER] Distro: [$currentDistro] homeDir[$HOME] Uid: [$UID]===---"
+
 if [[ "$USER" == "root" ]]
 then
     if [[ "$currentDistro" == "debian os" ]] || [[ "$currentDistro" == "debian ec2 os" ]]
     then
         log info "On est sur une distro [$currentDistro] pour le [$USER]"
         createConfigureSshRoot
+        copyGitBashCompletion
+        copyGitBashPrompt
         copyRootBashrc
         copyShAliases
         copyVimrc
@@ -153,19 +267,49 @@ then
         log info "On est sur une distro [$currentDistro] pour le [$USER] UNKOWNW"
     fi
 
+    if [[ "$userUid1000" == "jenkins" ]] && detectIfInContainer
+    then
+      insertPS1BashBashrc
+    fi
+
 elif [[ "$USER" == "vagrant" ]]
 then
 
     if [[ "$currentDistro" == "debian os" ]] || [[ "$currentDistro" == "debian ec2 os" ]]
     then
+        copyGitBashCompletion
+        copyGitBashPrompt
         copyUserBashrc
+        copyUserBashAliases
         copyShAliases
         copyVimrc
         copyPsqlRc
     else
         log info "On est sur une distro [$currentDistro] pour le [$USER] UNKONW"
     fi
+else
+    if [[ "$currentDistro" == "debian os" ]] || [[ "$currentDistro" == "debian ec2 os" ]]
+    then
+        copyGitBashPrompt
+        copyUserBashrc
+        copyUserBashAliases
+        copyShAliases
+        copyVimrc
+        copyPsqlRc
+#    else
+#        log info "On est sur une distro [$currentDistro] pour le [$USER] UNKONW"
+    fi
+
+    if detectIfInContainer && [[ "$USER" == "jenkins" ]]
+    then
+      echo "on est dans un container JENKINS ON on agit en consequence"
+      echo "PS1='${debian_chroot:+($debian_chroot)}\[\033[01;31m\]\u@\h\[\033[00m\] [\D{%T}] \[\033[01;34m\]\w\[\033[00m\]\[\033[36;40m\]\[\033[00m\] > '" >> /etc/bash.bashrc
+
+    fi
+
 fi
+
+
 
 cleanup
 log info "---=== Fin de Configuration de becane ===---"
